Follow advice of @SFulpius for better structured tests
