+++ app/src/androidTest/java/ch/epfl/sweng/swenggolf/ShowOfferActivityDbNotWorkingTest.java
@@ -18,7 +18,6 @@ import ch.epfl.sweng.swenggolf.tools.FragmentConverter;
 
 import static android.support.test.espresso.Espresso.onView;
 import static android.support.test.espresso.action.ViewActions.click;
-import static android.support.test.espresso.action.ViewActions.scrollTo;
 import static android.support.test.espresso.assertion.ViewAssertions.matches;
 import static android.support.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static android.support.test.espresso.matcher.ViewMatchers.withId;
@@ -60,7 +59,7 @@ public class ShowOfferActivityDbNotWorkingTest {
     @Test
     public void displaysErrorMessageWhenAnswersCannotBeLoaded() {
         database.setEntryNotWorking(Database.ANSWERS_PATH, offer.getUuid());
-        AnswersTest.showOfferCustomScrollTo();
+        TestUtility.showOfferCustomScrollTo();
         onView(withId(R.id.error_message)).check(matches(isDisplayed()));
     }
